+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Permalink to each world record on the homepage list
 - Page to show each world record
+- Number of submissions waiting in the menu bar
 
 ### Changed